refactor: rename cmp
