+++ frontend/src/lib/components/accounts/IcrcWalletPage.svelte
